refactor: move presets logic to presets.py
